% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_graph.R
\name{geomgraph}
\alias{geomgraph}
\title{Geometric Graph}
\usage{
geomgraph(x, from, to, r)
}
\arguments{
\item{x}{matrix}

\item{from}{indices}

\item{to}{indices}

\item{r}{radius}
}
\description{
Geometric Graph
}
