% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_ApproximationAlphaClassify_alt.R
\name{MCMC_ApproximationAlphaClassify_alt}
\alias{MCMC_ApproximationAlphaClassify_alt}
\title{compute classification vector  MCMC with auxiliary method 1 means Strauss point 0 Poisson point}
\usage{
MCMC_ApproximationAlphaClassify_alt(x, r, mcmcPar)
}
\arguments{
\item{r}{hardcore radius of the Strauss process}

\item{datappE}{simulated superposition, NEED SIMULATION IN THE WINDOW [-r,1+r]^2}
}
\description{
compute classification vector  MCMC with auxiliary method 1 means Strauss point 0 Poisson point
}
