% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_classificationFromPosteriorsFinal.R
\name{MCMC_classificationFromPosteriors}
\alias{MCMC_classificationFromPosteriors}
\title{classification from posteriors and some subproducts}
\usage{
MCMC_classificationFromPosteriors(
  p,
  perc,
  DistancesBetweenPointsE,
  nrcloseNeighbors,
  priorExpectationPercentageNoise,
  r
)
}
\arguments{
\item{p}{posterior probability of points to be Strauss points}

\item{perc}{vector of probabilities to give to T-T T-F and F-F in couples}

\item{DistancesBetweenPointsE}{ditances between points of superposition pattern}

\item{nrcloseNeighbors, }{number of r-close neighbors for every point}

\item{priorExpectationPercentageNoise}{[TODO]}
}
\description{
classification from posteriors and some subproducts
}
