% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_computeClusters.R
\name{MCMC_computeClusters}
\alias{MCMC_computeClusters}
\title{compute indices of clusters in which to divide the posterior probabilities}
\usage{
MCMC_computeClusters(p, nrcloseNeighbors)
}
\arguments{
\item{p}{posterior probabilities to be a true bubble computed by MCMC}

\item{nrcloseNeighbors}{vector whose components contains the number of r-close neighbors of the points}
}
\description{
compute indices of clusters in which to divide the posterior probabilities
}
