% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_computePercentage.R
\name{MCMC_computePercentage}
\alias{MCMC_computePercentage}
\title{percentages of TT TF FF in r-close couples of a Strauss process with parameters theta}
\usage{
MCMC_computePercentage(theta, bbox)
}
\arguments{
\item{theta}{vector of parameters of the process}

\item{win}{window of observation}
}
\description{
percentages of TT TF FF in r-close couples of a Strauss process with parameters theta
}
