% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_computePosteriorsApproximationAlpha_alt.R
\name{MCMC_computePosteriorsApproximationAlpha_alt}
\alias{MCMC_computePosteriorsApproximationAlpha_alt}
\title{compute posteriors and estimation theta}
\usage{
MCMC_computePosteriorsApproximationAlpha_alt(r, x, mcmcPar)
}
\arguments{
\item{r}{radius of hardcore process}

\item{reps}{total number of iterations of MCMC algorithm}

\item{repin}{burn-in iterations of MCMC algorithm}

\item{k}{gap between considered iterations in MCMC algorithm}

\item{DistancesBetweenPointsE}{matrix of the distances between points of the superposition}

\item{indConsidered}{subindices of points which belong to the restricted window}

\item{binConsidered}{vector of 0 and 1, 1 means that the point is inside the restricted window}

\item{volA}{volume of restricted window}

\item{volAE}{volume of big window}

\item{bbox}{restricted window in matrix form}

\item{supportPriorLeft}{vector of left limits prior support}

\item{supportPriorRight}{vector of right limits prior support}
}
\description{
compute posteriors and estimation theta
}
