% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_computePosteriorsGaussianPriors.R
\name{MCMC_computePosteriorsGaussianPriors}
\alias{MCMC_computePosteriorsGaussianPriors}
\title{compute posteriors and estimation theta}
\usage{
MCMC_computePosteriorsGaussianPriors(r, reps, repin, k, DistancesBetweenPointsE,
  indConsidered, binConsidered, totE, volA, volAE, bbox, meanPriors, varPriors,
  tau, eps)
}
\arguments{
\item{r}{radius of hardcore process}

\item{reps}{total number of iterations of MCMC algorithm}

\item{repin}{burn-in iterations of MCMC algorithm}

\item{k}{gap between considered iterations in MCMC algorithm}

\item{DistancesBetweenPointsE}{matrix of the distances between points of the superposition}

\item{indConsidered}{subindices of points which belong to the restricted window}

\item{binConsidered}{vector of 0 and 1, 1 means that the point is inside the restricted window}

\item{volA}{volume of restricted window}

\item{volAE}{volume of big window}

\item{bbox}{restricted window in matrix form}

\item{meanPriors}{mean of the normal in the lognormal prior distributions}

\item{varPriors}{variance of the normal in the lognormal prior distributions}
}
\description{
compute posteriors and estimation theta
}
