% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_DensityConfigurationGivenZ.R
\name{MCMC_DensityConfigurationGivenZ}
\alias{MCMC_DensityConfigurationGivenZ}
\title{computes p(y0,n0,y1,n1|theta)}
\usage{
MCMC_DensityConfigurationGivenZ(ZE, theta, volAE, DistancesBetweenPointsE)
}
\arguments{
\item{theta}{current vector of parameters}

\item{Z}{vector of 0 and 1, labelled with 1 are Strauss points}

\item{ZrealE}{vector of 0 and 1, labelled with 1 are Strauss points in full window}

\item{volA}{volA volume of restricted window}

\item{DistancesBetweenPointE}{matrix with distances between points}
}
\description{
computes p(y0,n0,y1,n1|theta)
}
