% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_Init.R
\name{MCMC_InitZ}
\alias{MCMC_InitZ}
\title{Initialization of Z}
\usage{
MCMC_InitZ(nE, theta, lambda_1, volAE, eps, DistancesBetweenPointsE)
}
\arguments{
\item{nE}{total number of points}

\item{theta}{current vector of parameters}

\item{lambda_1}{current estiation of the intensity of the Strauss process}

\item{volAE}{volAE volume of full window}

\item{eps}{esp acceptance parameter for a configuration}

\item{DistancesBetweenPointsE}{matrix with distances between points}
}
\description{
Initialization of Z
}
