% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_RClose.R
\name{MCMC_RClose}
\alias{MCMC_RClose}
\title{Number of r-close couples in the realization of the Strauss process}
\usage{
MCMC_RClose(Z, r, DistancesBetweenPoints)
}
\arguments{
\item{Z}{vector of 0 and 1, labelled with 1 are Strauss points}

\item{r}{r interaction radius}

\item{DistancesBetweenPoints}{matrix with distances between points}
}
\description{
Number of r-close couples in the realization of the Strauss process
}
