% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_RCloseE.R
\name{MCMC_RCloseE}
\alias{MCMC_RCloseE}
\title{Number of r-close couples in the realization of the Strauss process where at least one point is in the restricted window}
\usage{
MCMC_RCloseE(Z, binConsidered, r, DistancesBetweenPoints)
}
\arguments{
\item{Z}{vector of 0 and 1, labelled with 1 are Strauss points}

\item{binConsidered}{vector of 0 and 1, labelled with 1 are the points in the restricted window}

\item{r}{r interaction radius}

\item{DistancesBetweenPoints}{matrix with distances between all points}
}
\description{
Number of r-close couples in the realization of the Strauss process where at least one point is in the restricted window
}
