% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_RCloseNeighbours.R
\name{MCMC_RCloseNeighbours}
\alias{MCMC_RCloseNeighbours}
\title{number of r-close neighbours of the m-th  point in the Strauss process}
\usage{
MCMC_RCloseNeighbours(Z, r, m, DistancesBetweenPoints)
}
\arguments{
\item{Z}{vector of 0 and 1, labelled with 1 are Strauss points}

\item{r}{interaction radius}

\item{m}{m identifies one particular point (the m-th point)}

\item{DistancesBetweenPoints}{matrix with distances between points}
}
\description{
number of r-close neighbours of the m-th  point in the Strauss process
}
