% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_UpdateThetaApproximationAlpha_alt.R
\name{MCMC_UpdateThetaApproximationAlpha_alt}
\alias{MCMC_UpdateThetaApproximationAlpha_alt}
\title{Update of theta in MCMC algorithm}
\usage{
MCMC_UpdateThetaApproximationAlpha_alt(
  ZE,
  theta,
  tau,
  supportPriorLeft,
  supportPriorRight,
  n0,
  n1,
  bbox,
  binConsidered,
  Phi,
  StG,
  lambda_1,
  volA
)
}
\arguments{
\item{ZE}{vector of classification in the big window}

\item{theta}{vector of actual parameters}

\item{tau}{variances for determine new proposals of parameters}

\item{supportPriorLeft}{left limits of prior support}

\item{supportPriorRight}{right limits of prior support}

\item{n0}{total number of Poisson points}

\item{n1}{total number of Strauss points}

\item{bbox}{retricted window in matrix form}

\item{binConsidered}{vector of 0 and 1, 1 means that the point is inside the restricted window}

\item{StG}{parameter for estimation of lambda_1}

\item{lambda_1}{current estiation of the intensity of the Strauss process}

\item{volA}{volume of restricted window}

\item{DistancesBetweenPointsE}{matrix with distances between points}
}
\description{
Update of theta in MCMC algorithm
}
