% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_UpdateThetaGaussianPriors.R
\name{MCMC_UpdateThetaGaussianPriors}
\alias{MCMC_UpdateThetaGaussianPriors}
\title{Update of theta in MCMC algorithm}
\usage{
MCMC_UpdateThetaGaussianPriors(
  ZE,
  theta,
  tau,
  meanPriors,
  varPriors,
  n0,
  n1,
  bbox,
  binConsidered,
  DistancesBetweenPointsE,
  StG,
  lambda_1,
  volA
)
}
\arguments{
\item{ZE}{vector of classification in the big window}

\item{theta}{vector of actual parameters}

\item{tau}{variances for determine new proposals of parameters}

\item{meanPriors}{mean of the prior lognormal distribution of the parameters}

\item{varPriors}{variance of the prior lognormal distribution of the parameters}

\item{n0}{total number of Poisson points}

\item{n1}{total number of Strauss points}

\item{bbox}{retricted window in matrix form}

\item{binConsidered}{vector of 0 and 1, 1 means that the point is inside the restricted window}

\item{DistancesBetweenPointsE}{matrix with distances between points}

\item{StG}{parameter for estimation of lambda_1}

\item{lambda_1}{current estiation of the intensity of the Strauss process}

\item{volA}{volume of restricted window}
}
\description{
Update of theta in MCMC algorithm
}
