% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMC_UpdateZ_alt.R
\name{MCMC_UpdateZ_alt}
\alias{MCMC_UpdateZ_alt}
\title{update Z}
\usage{
MCMC_UpdateZ_alt(Z, theta, n, Phi, lambda_1)
}
\arguments{
\item{Z}{current vector of classification}

\item{theta}{current vector of parameters}

\item{n}{n total number of points in the restricted window}

\item{lambda_1}{estiamtion of intensity of the Strauss process}

\item{DistancesBetweenPoints}{matrix with distances between points}
}
\description{
update Z
}
