% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCClassifyGaussianPriors.R
\name{MCMCClassifyGaussianPriors}
\alias{MCMCClassifyGaussianPriors}
\title{compute classification vector  MCMC with auxiliary method 1 means Strauss point 0 Poisson point}
\usage{
MCMCClassifyGaussianPriors(x, r, prior, reps, repin, k)
}
\arguments{
\item{x}{list  x=pattern to classify bbox=observation window}

\item{r}{hardcore radius of the Strauss process}

\item{prior}{information about parameters priors}
}
\description{
compute classification vector  MCMC with auxiliary method 1 means Strauss point 0 Poisson point
}
