% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCClassifyGaussianPriors.R
\name{MCMCClassifyGaussianPriors_Old}
\alias{MCMCClassifyGaussianPriors_Old}
\title{Check pattern  
Check priors}
\usage{
MCMCClassifyGaussianPriors_Old(x, r, prior)
}
\description{
Check pattern  
Check priors
}
