% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MCMCClassifyGaussianPriors.R
\name{MCMCClassifyGaussianPriorsFixedTheta}
\alias{MCMCClassifyGaussianPriorsFixedTheta}
\title{Done!}
\usage{
MCMCClassifyGaussianPriorsFixedTheta(x, theta, lambda_1, reps, repin, k)
}
\description{
Done!
}
