% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_graph.R
\name{nl2spam}
\alias{nl2spam}
\title{Transform neighbourhood list to a Matrix}
\usage{
nl2spam(x)
}
\arguments{
\item{x}{Neighbourhood list as given by \code{\link{geom}}}
}
\description{
Transform neighbourhood list to a Matrix
}
