% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_result_methods.R
\name{print.VBclassification}
\alias{print.VBclassification}
\title{Print method for VB classification}
\usage{
\method{print}{VBclassification}(x, ...)
}
\arguments{
\item{x}{object from VBclassify}
}
\description{
Print method for VB classification
}
