% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_check_pattern.R
\name{VB_check_pattern}
\alias{VB_check_pattern}
\title{Check input pattern}
\usage{
VB_check_pattern(x)
}
\arguments{
\item{x}{Point pattern candidate.}
}
\description{
Check input pattern
}
\details{
We use the simple format of a list with
$x = Matrix of size n x d, where d is dimension.
$bbox = 2 x d matrix, with ranges of bounding box in columns.
}
