% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_check_prior.R
\name{VB_check_prior}
\alias{VB_check_prior}
\title{Check priors for VBclassify}
\usage{
VB_check_prior(prior, x, ninja = FALSE)
}
\arguments{
\item{prior}{list of priors}

\item{x}{point pattern, list with $x and $bbox}
}
\description{
Check priors for VBclassify
}
\details{
Priors:
p0 ~ scalar[0,1], prior for P(z_i=1) for all i. Default: 0.5
m  ~ numeric(3),  mean for theta=log(lambda,beta,gamma) 
S  ~ Covariance matrix fof theta

Default prior variance is 10000, flat priors.
}
