% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_initialize.R
\name{VB_initial_state}
\alias{VB_initial_state}
\title{Create the initial state object for VB strauss}
\usage{
VB_initial_state(x, R, prior, rho, eps, quads, keep.Q = FALSE,
  ninja = FALSE)
}
\arguments{
\item{quads}{number of quads, shortest edge. Longer edge count according to aspect ratio.

spam version}
}
\description{
Create the initial state object for VB strauss
}
