% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VB_result_methods.R
\name{print.VBclassification_summary}
\alias{print.VBclassification_summary}
\title{Print the summary of VB classification result}
\usage{
\method{print}{VBclassification_summary}(x, ...)
}
\arguments{
\item{x}{VB classification result}
}
\description{
Print the summary of VB classification result
}
